refactor : clickDelete 함수를 events 폴더 안에 정리
