+++ src/events/clickDelete.js
@@ -3,9 +3,7 @@ import renderFilteredRestaurants from "../ui/renderFilteredRestaurant";
 
 const clickDelete = (restaurant, restaurantList) => {
   restaurantList.delete(restaurant);
-
   Persistence.saveRestaurantList(restaurantList.value);
-
   renderFilteredRestaurants(restaurantList);
 };
