+++ src/ui/changeModalContents.js
@@ -1,6 +1,6 @@
 import RestaurantCard from "../components/RestaurantCard/index.js";
 import RestaurantDetailButtonContainer from "../components/RestaurantDetailButtonContainer/index.js";
-import clickDelete from "../service/clickDelete.js";
+import clickDelete from "../events/clickDelete.js";
 import { $ } from "../utils/dom.js";
 import eventHandlers from "../events/eventHandlers.js";
